Move peak and averages to separate file
